forward

+++ src/apue/chapter5/file_lock.cc
@@ -67,7 +67,7 @@ private:
 void* WriteFile(void *arg) {
     FileLock* lock = static_cast<FileLock*>(arg);
     lock->Lock();
-    printf("process %ld lock file\n", getpid());
+    printf("process %ld lock file\n", (long)getpid());
     fflush(NULL);
     sleep(3);
     lock->Unlock();
@@ -76,7 +76,7 @@ void* WriteFile(void *arg) {
 
 int main() {
     FileLock lock;
-    printf("main process %ld\n", getpid());
+    printf("main process %ld\n", (long)getpid());
     pid_t pid;
     if ((pid = fork()) == 0) {
         // child processs
